Context class
Added icon font and made the code a bit better

+++ Tin/include/Tin/Renderer/Objects/Camera.hpp
@@ -11,23 +11,23 @@ namespace Tin {
 		glm::vec3 Position, Orientation;
 		float FOV, NearPlane, FarPlane;
 
-		Camera(glm::vec2 viewportSize, glm::vec3 position, glm::vec3 orientation, float fov = 90.0f, float nearPlane = 0.1f, float farPlane = 1000.0f);
+		Camera(glm::ivec2 viewportSize, glm::vec3 position, glm::vec3 orientation, float fov = 90.0f, float nearPlane = 0.1f, float farPlane = 1000.0f);
 
 		// Updates the projection and view matrix
 		// Uniforms required: CamProjection (mat4), CamView (mat4)
 		void UpdateMatrix(Shader& shader, glm::mat4 projectionMatrix, glm::mat4 viewMatrix) const;
 
 		// Sets the camera viewport size
-		void SetViewportSize(glm::vec2 size);
+		void SetViewportSize(glm::ivec2 size);
 
 		// Returns the projection matrix
 		glm::mat4 GetProjectionMatrix() const;
 		// Returns the view matrix
 		glm::mat4 GetViewMatrix() const;
 		// Returns the camera viewport size
-		glm::vec2 GetViewportSize() const;
+		glm::ivec2 GetViewportSize() const;
 	private:
-		glm::vec2 m_viewportSize;
+		glm::ivec2 m_viewportSize;
 	};
 }
